mal-bud-backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
av==12.3.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.4.0
distro==1.9.0
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
einops==0.8.0
filelock==3.16.1
flash_attn @ file:///C:/Users/Runner/Downloads/flash_attn-2.6.3%2Bcu123torch2.4.0cxx11abiFALSE-cp312-cp312-win_amd64.whl#sha256=7730828bcee150641a555ee98d6bfbe063be6d8278de38a3884ac6dc14a1097b
flatbuffers==24.3.25
fsspec==2024.9.0
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
onnxruntime==1.19.2
packaging==24.1
protobuf==5.28.2
pydantic==2.9.2
pydantic_core==2.23.4
pyreadline3==3.5.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
sniffio==1.3.1
sqlparse==0.5.1
sympy==1.13.3
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
